Seperate out raster utils for reuse elsewhere
